Portfolio/requirements.txt
Flask==3.0.0
Flask-Login==0.6.3
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
python-dotenv==1.0.1
SQLAlchemy==2.0.29
Werkzeug==3.0.1
python-slugify==8.0.4
Jinja2==3.1.3
Flask-WTF==1.2.1
google-generativeai==0.5.4